Move audio annotation to the annotation directory
