+++ frontend/src/components/annotation/audio_annotation/audio_hotkeys.vue
