+++ frontend/src/components/annotation/audio_annotation/render_elements/waveform_selector.vue
